SEDONA-713 Add docs.
